safari-534.53-branch/LayoutTests/sputnik/Conformance/12_Statement/12.9_The_return_Statement/S12.9_A1_T2-expected.txt

CONSOLE MESSAGE: line 79: SyntaxError: Parse error
S12.9_A1_T2

PASS Expected parsing failure

TEST COMPLETE
